Se crea un reclamo!

+++ Biblioteca Usuarios/Serializador.cs
@@ -56,8 +56,15 @@ namespace Biblioteca_Usuarios
         {
             try
             {
-                string json = JsonConvert.SerializeObject(dictionaryAdmins, Newtonsoft.Json.Formatting.Indented);
-                File.WriteAllText(path, json);
+                if (File.Exists(path))
+                {
+                    string json = JsonConvert.SerializeObject(dictionaryAdmins, Newtonsoft.Json.Formatting.Indented);
+                    File.WriteAllText(path, json);
+                }
+                else
+                {
+                    Console.WriteLine("El archivo JSON no existe.");
+                }
             }
             catch (Exception ex)
             {
@@ -81,7 +88,7 @@ namespace Biblioteca_Usuarios
             return dictionary;
         }
 
-        public static void WriteJson(string ruta, List<Tramites> lista)
+        public static void WriteJsonTramites(string ruta, List<Tramites> lista)
         {
             try
             {
